requirements.txt
requests>=2.31.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.2
pandas>=2.1.0
wordcloud>=1.9.2
jieba>=0.42.1
matplotlib>=3.8.0
numpy>=1.24.0
pillow>=10.0.0
lxml>=4.9.0

